refactor: move oversized inline test modules into sibling files
The unit tests beside the code had grown to half the length of the modules
they cover, pushing several files past two thousand lines. Each test module
moves to its own file under the module's directory, unchanged: the code reads
without scrolling past a test suite, and the tests still see the private items
they exercise.
